Move the websocket download handler to a public module
