+++ scrapy/core/downloader/handlers/websocket.py
