+++ scrapy/settings/default_settings.py
@@ -292,8 +292,8 @@ DOWNLOAD_HANDLERS_BASE = {
     "https": "scrapy.core.downloader.handlers.http11.HTTP11DownloadHandler",
     "s3": "scrapy.core.downloader.handlers.s3.S3DownloadHandler",
     "ftp": "scrapy.core.downloader.handlers.ftp.FTPDownloadHandler",
-    "ws": "scrapy.core.downloader.handlers._websocket.WebSocketDownloadHandler",
-    "wss": "scrapy.core.downloader.handlers._websocket.WebSocketDownloadHandler",
+    "ws": "scrapy.core.downloader.handlers.websocket.WebSocketDownloadHandler",
+    "wss": "scrapy.core.downloader.handlers.websocket.WebSocketDownloadHandler",
 }
 
 DOWNLOAD_MAXSIZE = 1024 * 1024 * 1024  # 1024m
